requirements.txt
vllm
openai
requests
mlflow
pandas
scikit-learn
python-dotenv
